--- lib/inc/vk_buffer.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#define VMA_STATIC_VULKAN_FUNCTIONS 1
-#define VMA_DYNAMIC_VULKAN_FUNCTIONS 0
-#define VMA_IMPLEMENTATION
-#include <vk_mem_alloc.h>
-#include <vulkan/vulkan.h>
-
-namespace vkrt {
-	
-
-
-} // namespace vkrt
-
